chore: Fix Drive Folders  Issue + Refactor Google storage strategy + Rename  DidManager class to CredentialEngine + move saveToGoogleDrive to utils/ directory
