+++ package.json
@@ -20,12 +20,14 @@
 		"@digitalbazaar/ed25519-verification-key-2020": "^4.1.0",
 		"@digitalbazaar/vc": "^6.3.0",
 		"@types/express": "^4.17.21",
+		"@types/uuid": "^10.0.0",
 		"did-method-key": "^0.7.0",
 		"express": "^4.19.2",
 		"fs-extra": "^11.2.0",
 		"jest": "^29.7.0",
 		"ts-jest": "^29.1.2",
-		"tsc": "^2.0.4"
+		"tsc": "^2.0.4",
+		"uuid": "^10.0.0"
 	},
 	"devDependencies": {
 		"@types/fs-extra": "^11.0.4",
