+++ src/types/decalration.d.ts
