+++ tsconfig.json
@@ -5,7 +5,7 @@
 		"moduleResolution": "node",
 		"outDir": "./dist",
 		"rootDir": "./src",
-		// "strict": true,
+		"strict": false,
 		"esModuleInterop": true,
 		"skipLibCheck": true
 	},
